feat: refactor sector management routes; implement joinSector functionality and remove sitAtSector route
