+++ backend/src/routes/ru/ru.routes.ts
@@ -1,6 +1,6 @@
 import auth from '../../middleware/auth.js';
 import { Router } from 'express';
-import { getMenus, getApiDoc, getSectors, sitAtSector } from './ru.controller.js';
+import { getMenus, getApiDoc, getSectors } from './ru.controller.js';
 const router = Router();
 
 router.get('/', getApiDoc);
@@ -8,7 +8,5 @@ router.get('/', getApiDoc);
 router.get('/:ruId/sectors', auth, getSectors);
 
 router.get('/menus', auth, getMenus);
-
-router.post('/sit', auth, sitAtSector);
 
 export default router;
